requirements.txt
#
####### requirements.txt with example helpers #######
#

###### Requirements without Version Specifiers ######
setuptools
pandas
aistac-foundation
discovery-connectors
matplotlib
numpy
# SQl and HTML libs
#sqlalchemy
#lxml
#html5lib
#BeautifulSoup4

#
###### Requirements with Version Specifiers ######
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers
# Example:
#   cocopt == 0.6.1             # Version Matching. Must be version 0.6.1
#   keyring >= 4.1.1            # Minimum version 4.1.1
#   coverage != 3.5             # Version Exclusion. Anything except version 3.5
#   Mopidy-Dirble ~= 1.1        # Compatible release. Same as >= 1.1, == 1.*



#
###### Refer to other requirements files ######
# Example:
#   -r other-requirements.txt



#
###### A particular file ######
# Example:
#   ./downloads/numpy-1.9.2-cp34-none-win32.whl
#   http://wxpython.org/Phoenix/snapshot-builds/wxPython_Phoenix-3.0.3.dev1820+49a8884-cp34-none-win_amd64.whl

